Added lots of documentation and comments to chat.py and test_achat.py
Renamed test_chat.py to test_achat.py so that it runs first in
the test suite. Otherwise the test client always fails to connect.
I believe that this is an issue with the testing framework, but if
there are issues with WebSockets in the future, it may come back to this.